modules/backup/function/requirements.txt
google-cloud-storage>=2.0.0